Adding in a cleanup flag that can prevent cleanup which is useful for testing. Won't want to do this in production of course so the default is true, to auto-cleanup intermediate files and the final result after uploading to GNOS.

+++ src/main/java/com/github/seqware/WorkflowClient.java
@@ -31,6 +31,7 @@ public class WorkflowClient extends OicrWorkflow {
   boolean useGtDownload = true;
   boolean useGtUpload = true;
   boolean useGtValidation = true;
+  boolean cleanup = true;
   // number of splits for bam files, default 1=no split
   int bamSplits = 1;
   String reference_path = null;
@@ -112,7 +113,7 @@ public class WorkflowClient extends OicrWorkflow {
       if (getProperty("use_gtdownload") != null) { if("false".equals(getProperty("use_gtdownload"))) { useGtDownload = false; } }
       if (getProperty("use_gtupload") != null) { if("false".equals(getProperty("use_gtupload"))) { useGtUpload = false; } }
       if (getProperty("use_gtvalidation") != null) { if("false".equals(getProperty("use_gtvalidation"))) { useGtValidation = false; } }
-
+      if (getProperty("cleanup") != null) { if("false".equals(getProperty("cleanup"))) { cleanup = false; } }
       if (!getProperty("numOfThreads").isEmpty()) {
         numOfThreads = Integer.parseInt(getProperty("numOfThreads"));
       }
@@ -239,7 +240,7 @@ public class WorkflowClient extends OicrWorkflow {
         qcJobs.add(qcJob);
 
         // CLEANUP DOWNLOADED INPUT UNALIGNED BAM FILES
-        if (useGtDownload) {
+        if (useGtDownload && cleanup) {
           Job cleanup1 = this.getWorkflow().createBashJob("cleanup_" + i);
           cleanup1.getCommand().addArgument("rm -f " + file);
           cleanup1.setMaxMemory(smallJobMemM);
@@ -294,7 +295,7 @@ public class WorkflowClient extends OicrWorkflow {
         qcJobs.add(qcJob);
 
         // CLEANUP DOWNLOADED INPUT UNALIGNED BAM FILES
-        if (useGtDownload) {
+        if (useGtDownload && cleanup) {
           Job cleanup1 = this.getWorkflow().createBashJob("cleanup2_" + i);
           cleanup1.getCommand().addArgument("rm -f " + file);
           cleanup1.setMaxMemory(smallJobMemM);
@@ -427,12 +428,14 @@ public class WorkflowClient extends OicrWorkflow {
     
     
     // CLEANUP LANE LEVEL BAM FILES
-    for (int i = 0; i < numBamFiles; i++) {
-      Job cleanup2 = this.getWorkflow().createBashJob("cleanup3_" + i);
-      cleanup2.getCommand().addArgument("rm -f out_" + i + ".bam");
-      cleanup2.addParent(job04);
-      cleanup2.setMaxMemory(smallJobMemM);
-      cleanup2.addParent(qcJobs.get(i));
+    if (cleanup) {
+      for (int i = 0; i < numBamFiles; i++) {
+        Job cleanup2 = this.getWorkflow().createBashJob("cleanup3_" + i);
+        cleanup2.getCommand().addArgument("rm -f out_" + i + ".bam");
+        cleanup2.addParent(job04);
+        cleanup2.setMaxMemory(smallJobMemM);
+        cleanup2.addParent(qcJobs.get(i));
+      }
     }
 
     // PREPARE METADATA & UPLOAD
@@ -486,16 +489,18 @@ public class WorkflowClient extends OicrWorkflow {
     job06.addParent(mergeUnmappedJob);
 
     // CLEANUP FINAL BAM
-    Job cleanup3 = this.getWorkflow().createBashJob("cleanup3");
-    cleanup3.getCommand().addArgument("rm -f *.bam "
-                                       + this.dataDir + outputFileName
-                                       + " " + this.dataDir + outputUnmappedFileName); 
-    cleanup3.addParent(job05);
-    cleanup3.addParent(job06);
-    for (Job qcJob : qcJobs) {
-      cleanup3.addParent(qcJob);
+    if (cleanup) {
+      Job cleanup3 = this.getWorkflow().createBashJob("cleanup3");
+      cleanup3.getCommand().addArgument("rm -f *.bam "
+                                         + this.dataDir + outputFileName
+                                         + " " + this.dataDir + outputUnmappedFileName); 
+      cleanup3.addParent(job05);
+      cleanup3.addParent(job06);
+      for (Job qcJob : qcJobs) {
+        cleanup3.addParent(qcJob);
+      }
+      cleanup3.setMaxMemory(smallJobMemM);
     }
-    cleanup3.setMaxMemory(smallJobMemM);
 
   }
